requirements.txt
wechatpy>=1.8.0
openai>=1.12.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
requests>=2.31.0
